feat: Requests start in a more deterministic way (#98)
Previous implementation used `requestIdleCallback` to defer requests
being sent to avoid a race condition with cached requests / promise
chaining.  `requestIdleCallback` means that is it possible for a request
to be delayed for multiple seconds before being sent.

It it this none-deterministic behaviour which is making some tests
"flakey".  It doesn't seem like a good idea to keep this... instead the
request execution is delayed via a `setTimeout` call.